PharmaAssist/requirements.txt.txt
google-cloud-vision
google-auth
google-auth-oauthlib
pandas
opencv-python
rapidfuzz
numpy
